feat: scaffold sidebar chat WebviewView
