+++ package.json
@@ -26,7 +26,8 @@
     "onCommand:openhands.reconnect",
     "onCommand:openhands.pauseCurrentRun",
     "onCommand:openhands.resumeCurrentRun",
-    "onView:openhands.quickActions"
+    "onView:openhands.quickActions",
+    "onView:openhands.chat"
   ],
   "contributes": {
 
@@ -202,6 +203,11 @@
     },
     "views": {
       "openhands": [
+        {
+          "id": "openhands.chat",
+          "name": "Chat",
+          "type": "webview"
+        },
         {
           "id": "openhands.quickActions",
           "name": "OpenHands",
